Remove generated style upon destruction of objects
An IE "bug-feature" (up to Version 9) limits stylesheets to a total count of 31. Rich apps with many widgets can hit that limit easily.

view.stylize() adds a <style> tag to <head> for each instantiated object.

This patch adds an ID to this style tag and makes sure that it is removed from <head> when .destroy() is called.